Keeping it simple for now and do the dependency management in cmake.
CMakeLists.txt already handles the two OCOS_ENABLE_* values that need opencv with

```cmake
if(OCOS_ENABLE_CV2 OR OCOS_ENABLE_VISION)
  set(_ENABLE_OPENCV ON)
  message(STATUS "Fetch opencv")
  include(opencv)
endif()
```

and the vision ops are included separately to the cv2 ops so we don't bloat the mobile package due to opencv usage by the cv2 ops.

The other OCOS_ENABLE_* groups are 1:1 between ops and dependency.

If we get more overlapping things in the future we can re-consider if something more complicated is needed.